+++ src/org/zlibrary/ui/swing/application/ZLSwingApplicationWindow.java
@@ -40,8 +40,8 @@ public class ZLSwingApplicationWindow extends ZLApplicationWindow {
 	
 	private class ZLFrame extends JFrame {
 		ZLFrame() {
-			setSize((int)myWidthOption.getValue(), (int)myHeightOption.getValue());
-			setLocation((int)myXOption.getValue(), (int)myYOption.getValue());
+			setSize(myWidthOption.getValue(), myHeightOption.getValue());
+			setLocation(myXOption.getValue(), myYOption.getValue());
 			addComponentListener(new ComponentAdapter() {
 				public void componentResized(ComponentEvent event) {
 					if (!isFullscreen()) {
